backend/services/gpu_worker_client/requirements.txt
httpx>=0.27.0
redis>=5.0.3
pydantic>=2.6.3
python-dotenv>=1.0.1
torch>=2.2.0
transformers>=4.39.0
pynvml>=11.5.0
